// src/FX.d.ts
export interface IBoxCoreParams {
    width: number;
    height: number;
    radial: boolean;
}